implemented base BLE device package and move fragmenter
