+++ internal/ble/ble.go
@@ -1,104 +1,72 @@
 package ble
 
 import (
-	"sync"
-	"time"
-
 	"github.com/paypal/gatt"
+	"github.com/paypal/gatt/examples/option"
 )
 
-// Client has the methods used by the service when
-// characteristics are requested.
-type Client interface {
-	Write(data []byte) error
-	Notify() <-chan []byte
-}
+// State indicates the state of the connection with
+// the central
+type State byte
 
 const (
-	continuePacket = 0x00
-	endPacket      = 0x01
-	payloadSize    = 0x13
-	packetInterval = 100
+	name = "terminal"
 )
 
 var (
-	serviceUUID = gatt.UUID16(0xAAAA)
-	writeUUID   = gatt.UUID16(0xBBBB)
-	notifyUUID  = gatt.UUID16(0xCCCC)
+	// Connected represents the state when the central gets connected.
+	Connected State = 0x00
+	// Disconnected represents the state when the central gets disconnected.
+	Disconnected State = 0x01
 )
 
-type receiver struct {
-	mu      sync.RWMutex
-	lastRcv int64
-
-	ticker *time.Ticker
-	data   []byte
+// ConnListener has a method ConnState that it's called when
+// the state of the connection with the central changes.
+type ConnListener interface {
+	ConnState(state State)
 }
 
-func (r *receiver) received(data []byte) {
-	r.mu.Lock()
-	r.lastRcv = time.Now().Unix() * 1000
-	if len(r.data) == 0 {
-		r.ticker = time.NewTicker(1 * time.Second)
-		go r.check()
+// Device is the device that holds the connection to the central.
+type Device struct {
+	c            gatt.Central
+	connListener ConnListener
+}
+
+// Init initializes the BLE device.
+func Init(connListener ConnListener, services ...*gatt.Service) (*Device, error) {
+	d, err := gatt.NewDevice(option.DefaultServerOptions...)
+	if err != nil {
+		return nil, err
 	}
-	r.mu.Unlock()
-	r.data = append(r.data, data...)
-}
-
-func (r *receiver) check() {
-	for n := range r.ticker.C {
-		r.mu.RLock()
-		if (n.Unix()*1000)-r.lastRcv < 500 {
-			r.mu.RUnlock()
-			continue
+	dvc := &Device{connListener: connListener}
+	d.Handle(
+		gatt.CentralConnected(dvc.centralConnected),
+		gatt.CentralDisconnected(dvc.centralDisconnected),
+	)
+	onStateChanged := func(d gatt.Device, s gatt.State) {
+		switch s {
+		case gatt.StatePoweredOn:
+			uuids := []gatt.UUID{}
+			for _, svc := range services {
+				uuids = append(uuids, svc.UUID())
+			}
+			d.AdvertiseNameAndServices(name, uuids)
 		}
-		r.mu.RUnlock()
-		r.data = []byte{}
-		return
 	}
+	return dvc, d.Init(onStateChanged)
 }
 
-func (r *receiver) reset() {
-	r.data = []byte{}
-	r.ticker.Stop()
+func (d *Device) centralConnected(c gatt.Central) {
+	d.connListener.ConnState(Connected)
+	d.c = c
 }
 
-// NewService returns a new gatt service with the characteristics
-// of the IDK protocol.
-func NewService(client Client) *gatt.Service {
-	rcv := &receiver{data: []byte{}}
-	s := gatt.NewService(gatt.MustParseUUID(serviceUUID.String()))
-	s.AddCharacteristic(gatt.MustParseUUID(writeUUID.String())).HandleWriteFunc(
-		func(r gatt.Request, data []byte) (status byte) {
-			rcv.received(data[1:])
-			var s byte = gatt.StatusSuccess
-			if data[0] == continuePacket {
-				return s
-			}
-			if err := client.Write(rcv.data); err != nil {
-				s = gatt.StatusUnexpectedError
-			}
-			rcv.reset()
-			return s
-		})
-	s.AddCharacteristic(gatt.MustParseUUID(notifyUUID.String())).HandleNotifyFunc(
-		func(r gatt.Request, n gatt.Notifier) {
-			for data := range client.Notify() {
-				size := len(data)
-				for i := 0; i <= size/payloadSize; i++ {
-					b := []byte{continuePacket}
-					packetSize := payloadSize
-					if i == size/payloadSize {
-						b[0] = endPacket
-						packetSize = size - payloadSize*i
-					}
-					b = append(b, data[i*payloadSize:i*payloadSize+packetSize]...)
-					n.Write(b)
-					time.Sleep(packetInterval * time.Millisecond)
-				}
+func (d *Device) centralDisconnected(c gatt.Central) {
+	d.connListener.ConnState(Disconnected)
+	d.c = c
+}
 
-			}
-		})
-	return s
+// Close closes the connection to the central.
+func (d *Device) Close() error {
+	return d.c.Close()
 }
